requirements.txt
Flask==3.0.1
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
graphene==3.0.0b7
Flask-GraphQL==2.0.1
python-dotenv==1.0.0
